+++ tests/test_collections.py
@@ -119,7 +119,6 @@ def test_tilejson_collections(app):
     assert response.headers["content-type"] == "application/json"
     assert response.status_code == 200
     resp = response.json()
-    assert resp["name"] == f"Mosaic for '{collection_id}' Collection"
     assert resp["minzoom"] == 0
     assert resp["maxzoom"] == 24
     assert round(resp["bounds"][0]) == -180
@@ -249,36 +248,25 @@ def test_tiles_collections(rio, app):
 def test_wmts_collections(app):
     """Create wmts document."""
     # missing assets
-    response = app.get(
-        f"/collections/{collection_id}/WebMercatorQuad/WMTSCapabilities.xml"
-    )
+    response = app.get(f"/collections/{collection_id}/WMTSCapabilities.xml")
     assert response.status_code == 400
     assert (
         response.json()["detail"]
         == "Could not find any valid layers in metadata or construct one from Query Parameters."
     )
 
-    response = app.get(
-        f"/collections/{collection_id}/WebMercatorQuad/WMTSCapabilities.xml?assets=cog"
-    )
+    response = app.get(f"/collections/{collection_id}/WMTSCapabilities.xml?assets=cog")
     assert response.status_code == 200
     assert response.headers["content-type"] == "application/xml"
 
     # Validate it's a good WMTS
     with rasterio.open(io.BytesIO(response.content)) as src:
-        assert src.crs == "epsg:3857"
         assert src.profile["driver"] == "WMTS"
-
-    response = app.get(
-        f"/collections/{collection_id}/WorldCRS84Quad/WMTSCapabilities.xml?assets=cog"
-    )
-    assert response.status_code == 200
-    assert response.headers["content-type"] == "application/xml"
-
-    # Validate it's a good WMTS
-    with rasterio.open(io.BytesIO(response.content)) as src:
-        assert src.crs == "OGC:CRS84"
-        assert src.profile["driver"] == "WMTS"
+        assert len(src.subdatasets) == 13
+        with rasterio.open(
+            io.BytesIO(response.content), layer="default_WebMercatorQuad"
+        ) as sds:
+            assert sds.crs == "epsg:3857"
 
 
 @patch("rio_tiler.io.rasterio.rasterio")
@@ -520,14 +508,14 @@ def test_query_point_collections(app):
     response = app.get(
         f"/collections/{collection_id}/point/-85.5,36.1624", params={"assets": "cog"}
     )
-
     assert response.status_code == 200
     resp = response.json()
-    values = resp["values"]
+    values = resp["assets"]
     assert len(values) == 2
-    assert values[0][0] == "noaa-emergency-response/20200307aC0853130w361030"
-    assert values[0][2] == ["cog_b1", "cog_b2", "cog_b3"]
-    assert values[1][0] == "noaa-emergency-response/20200307aC0853000w361030"
+    assert values[0]["name"] == "noaa-emergency-response/20200307aC0853130w361030"
+    assert values[0]["band_names"] == ["cog_b1", "cog_b2", "cog_b3"]
+    assert values[0]["values"] == [27.0, 34.0, 42.0]
+    assert values[1]["name"] == "noaa-emergency-response/20200307aC0853000w361030"
 
     # with coord-crs
     response = app.get(
@@ -536,7 +524,7 @@ def test_query_point_collections(app):
     )
     assert response.status_code == 200
     resp = response.json()
-    assert len(resp["values"]) == 2
+    assert len(resp["assets"]) == 2
 
     # CollectionId not found
     response = app.get(
@@ -558,26 +546,26 @@ def test_query_point_collections(app):
 def test_collections_render(app, tmp_path):
     """Create wmts document."""
     response = app.get(
-        "/collections/MAXAR_BayofBengal_Cyclone_Mocha_May_23/WebMercatorQuad/WMTSCapabilities.xml"
+        "/collections/MAXAR_BayofBengal_Cyclone_Mocha_May_23/WMTSCapabilities.xml"
     )
     assert response.status_code == 200
     assert response.headers["content-type"] == "application/xml"
 
-    wmts = tmp_path / "WMTSCapabilities.xml"
-    with open(wmts, "wb") as f:
-        f.write(response.content)
-
     # Validate it's a good WMTS
-    with rasterio.open(wmts) as src:
+    with rasterio.open(io.BytesIO(response.content)) as src:
         assert not src.crs
         assert src.profile["driver"] == "WMTS"
-        assert len(src.subdatasets) == 3
-        assert ["color", "visual", "visualr"] == [
-            s.split(",layer=")[1] for s in src.subdatasets
-        ]
-        with rasterio.open(src.subdatasets[0]) as sub:
-            assert sub.crs == CRS.from_epsg(3857)
-            assert sub.profile["driver"] == "WMTS"
+        assert len(src.subdatasets) == 3 * 13  # 3 renders, 13 TMS
+        sds_names = [s.split(",layer=")[1] for s in src.subdatasets]
+        assert "color_WebMercatorQuad" in sds_names
+        assert "visualr_WebMercatorQuad" in sds_names
+        assert "color_WebMercatorQuad" in sds_names
+
+        with rasterio.open(
+            io.BytesIO(response.content), layer="color_WebMercatorQuad"
+        ) as sds:
+            assert sds.crs == CRS.from_epsg(3857)
+            assert sds.profile["driver"] == "WMTS"
 
     response = app.get("/collections/MAXAR_BayofBengal_Cyclone_Mocha_May_23/info")
     assert response.status_code == 200